requirements.txt
pyrogram==2.0.106
TgCrypto==1.2.5
telethon==1.34.0
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.0
asyncio
